+++ packages/matter-node.js/test/interaction/InteractionProtocolTest.ts
@@ -4,13 +4,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { Crypto } from "@project-chip/matter.js/crypto";
-import { KEY } from "../cluster/ClusterServerTestingUtil.js";
-
 import {
     AccessControlCluster,
     AccessLevel,
     AdministratorCommissioning,
+    BasicInformation,
     BasicInformationCluster,
     ClusterServer,
     ClusterServerObjForCluster,
@@ -18,6 +16,7 @@ import {
     WritableAttribute,
 } from "@project-chip/matter.js/cluster";
 import { Message, SessionType } from "@project-chip/matter.js/codec";
+import { Crypto } from "@project-chip/matter.js/crypto";
 import {
     AttributeId,
     ClusterId,
@@ -63,6 +62,7 @@ import {
 } from "@project-chip/matter.js/tlv";
 import { ByteArray } from "@project-chip/matter.js/util";
 import * as assert from "assert";
+import { KEY } from "../cluster/ClusterServerTestingUtil.js";
 
 const DummyTestDevice = DeviceTypeDefinition({
     code: 0,
@@ -72,7 +72,7 @@ const DummyTestDevice = DeviceTypeDefinition({
 });
 
 const READ_REQUEST: ReadRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     isFabricFiltered: true,
     attributeRequests: [
         { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), attributeId: AttributeId(2) },
@@ -84,10 +84,16 @@ const READ_REQUEST: ReadRequest = {
         { endpointId: undefined, clusterId: ClusterId(0x99), attributeId: AttributeId(3) }, // ignore
         { endpointId: EndpointNumber(0), clusterId: ClusterId(0x1d), attributeId: AttributeId(1) },
     ],
+    eventRequests: [
+        { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), eventId: EventId(0) }, // existing event
+        { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), eventId: EventId(400) }, // unsupported event
+        { endpointId: EndpointNumber(0), clusterId: ClusterId(0x99), eventId: EventId(4) }, // unsupported cluster
+        { endpointId: EndpointNumber(1), clusterId: ClusterId(0x28), eventId: EventId(1) }, // unsupported endpoint
+    ],
 };
 
 const READ_REQUEST_WITH_UNUSED_FILTER: ReadRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     isFabricFiltered: true,
     attributeRequests: [
         { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), attributeId: AttributeId(2) },
@@ -100,10 +106,17 @@ const READ_REQUEST_WITH_UNUSED_FILTER: ReadRequest = {
         { endpointId: EndpointNumber(0), clusterId: ClusterId(0x1d), attributeId: AttributeId(1) },
     ],
     dataVersionFilters: [{ path: { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28) }, dataVersion: 1 }],
+    eventRequests: [
+        { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), eventId: EventId(0) }, // existing event
+        { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), eventId: EventId(400) }, // unsupported event
+        { endpointId: EndpointNumber(0), clusterId: ClusterId(0x99), eventId: EventId(4) }, // unsupported cluster
+        { endpointId: EndpointNumber(1), clusterId: ClusterId(0x28), eventId: EventId(1) }, // unsupported endpoint
+    ],
+    eventFilters: [{ eventMin: 1 }],
 };
 
 const READ_REQUEST_WITH_FILTER: ReadRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     isFabricFiltered: true,
     attributeRequests: [
         { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), attributeId: AttributeId(2) },
@@ -116,12 +129,18 @@ const READ_REQUEST_WITH_FILTER: ReadRequest = {
         { endpointId: EndpointNumber(0), clusterId: ClusterId(0x1d), attributeId: AttributeId(1) },
     ],
     dataVersionFilters: [{ path: { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28) }, dataVersion: 0 }],
+    eventRequests: [
+        { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), eventId: EventId(0) }, // existing event
+        { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), eventId: EventId(400) }, // unsupported event
+        { endpointId: EndpointNumber(0), clusterId: ClusterId(0x99), eventId: EventId(4) }, // unsupported cluster
+        { endpointId: EndpointNumber(1), clusterId: ClusterId(0x28), eventId: EventId(1) }, // unsupported endpoint
+    ],
+    eventFilters: [{ eventMin: 2 }],
 };
 
 const READ_RESPONSE: DataReportPayload = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
-    eventReportsPayload: undefined,
     attributeReportsPayload: [
         {
             attributeData: {
@@ -178,12 +197,65 @@ const READ_RESPONSE: DataReportPayload = {
             },
         },
     ],
+    eventReportsPayload: [
+        {
+            eventData: {
+                path: {
+                    endpointId: EndpointNumber(0),
+                    clusterId: ClusterId(0x28),
+                    eventId: EventId(0),
+                    isUrgent: undefined,
+                },
+                schema: BasicInformation.TlvStartUpEvent,
+                payload: {
+                    softwareVersion: 1,
+                },
+                eventNumber: 1,
+                priority: 2,
+                epochTimestamp: 0,
+            },
+        },
+        {
+            eventData: {
+                path: {
+                    endpointId: EndpointNumber(0),
+                    clusterId: ClusterId(0x28),
+                    eventId: EventId(0),
+                    isUrgent: undefined,
+                },
+                schema: BasicInformation.TlvStartUpEvent,
+                payload: {
+                    softwareVersion: 2,
+                },
+                eventNumber: 2,
+                priority: 2,
+                epochTimestamp: 0,
+            },
+        },
+        {
+            eventStatus: {
+                path: { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), eventId: EventId(400) },
+                status: { status: 199 },
+            },
+        },
+        {
+            eventStatus: {
+                path: { endpointId: EndpointNumber(0), clusterId: ClusterId(0x99), eventId: EventId(4) },
+                status: { status: 195 },
+            },
+        },
+        {
+            eventStatus: {
+                path: { endpointId: EndpointNumber(1), clusterId: ClusterId(0x28), eventId: EventId(1) },
+                status: { status: 127 },
+            },
+        },
+    ],
 };
 
 const READ_RESPONSE_WITH_FILTER: DataReportPayload = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
-    eventReportsPayload: undefined,
     attributeReportsPayload: [
         {
             attributeStatus: {
@@ -216,10 +288,47 @@ const READ_RESPONSE_WITH_FILTER: DataReportPayload = {
             },
         },
     ],
+    eventReportsPayload: [
+        {
+            eventData: {
+                path: {
+                    endpointId: EndpointNumber(0),
+                    clusterId: ClusterId(0x28),
+                    eventId: EventId(0),
+                    isUrgent: undefined,
+                },
+                schema: BasicInformation.TlvStartUpEvent,
+                payload: {
+                    softwareVersion: 2,
+                },
+                eventNumber: 2,
+                priority: 2,
+                epochTimestamp: 0,
+            },
+        },
+        {
+            eventStatus: {
+                path: { endpointId: EndpointNumber(0), clusterId: ClusterId(0x28), eventId: EventId(400) },
+                status: { status: 199 },
+            },
+        },
+        {
+            eventStatus: {
+                path: { endpointId: EndpointNumber(0), clusterId: ClusterId(0x99), eventId: EventId(4) },
+                status: { status: 195 },
+            },
+        },
+        {
+            eventStatus: {
+                path: { endpointId: EndpointNumber(1), clusterId: ClusterId(0x28), eventId: EventId(1) },
+                status: { status: 127 },
+            },
+        },
+    ],
 };
 
 const INVALID_SUBSCRIBE_REQUEST: SubscribeRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     isFabricFiltered: true,
     attributeRequests: [
         { endpointId: EndpointNumber(0), clusterId: ClusterId(0x99), attributeId: AttributeId(2) },
@@ -235,7 +344,7 @@ const INVALID_SUBSCRIBE_REQUEST: SubscribeRequest = {
 };
 
 const WRITE_REQUEST: WriteRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     timedRequest: false,
     writeRequests: [
@@ -272,7 +381,7 @@ const WRITE_REQUEST: WriteRequest = {
 };
 
 const WRITE_RESPONSE: WriteResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     writeResponses: [
         {
             path: { attributeId: AttributeId(100), clusterId: ClusterId(40), endpointId: EndpointNumber(0) },
@@ -302,7 +411,7 @@ const WRITE_RESPONSE: WriteResponse = {
 };
 
 const WRITE_REQUEST_TIMED_REQUIRED: WriteRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     timedRequest: false,
     writeRequests: [
@@ -316,7 +425,7 @@ const WRITE_REQUEST_TIMED_REQUIRED: WriteRequest = {
 };
 
 const WRITE_RESPONSE_TIMED_REQUIRED: WriteResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     writeResponses: [
         {
             path: { attributeId: AttributeId(5), clusterId: ClusterId(40), endpointId: EndpointNumber(0) },
@@ -326,7 +435,7 @@ const WRITE_RESPONSE_TIMED_REQUIRED: WriteResponse = {
 };
 
 const WRITE_RESPONSE_TIMED_ERROR: WriteResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     writeResponses: [
         {
             path: { attributeId: AttributeId(5), clusterId: ClusterId(40), endpointId: EndpointNumber(0) },
@@ -336,7 +445,7 @@ const WRITE_RESPONSE_TIMED_ERROR: WriteResponse = {
 };
 
 const MASS_WRITE_REQUEST: WriteRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     timedRequest: false,
     writeRequests: [
@@ -360,7 +469,7 @@ const MASS_WRITE_REQUEST: WriteRequest = {
 };
 
 const MASS_WRITE_RESPONSE: WriteResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     writeResponses: [
         {
             path: { attributeId: AttributeId(5), clusterId: ClusterId(40), endpointId: EndpointNumber(0) },
@@ -386,7 +495,7 @@ const TlvAclTestSchema = TlvObject({
 });
 
 const CHUNKED_ARRAY_WRITE_REQUEST: WriteRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     timedRequest: false,
     writeRequests: [
@@ -447,7 +556,7 @@ const CHUNKED_ARRAY_WRITE_REQUEST: WriteRequest = {
 };
 
 const CHUNKED_ARRAY_WRITE_RESPONSE: WriteResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     writeResponses: [
         {
             path: { attributeId: AttributeId(0), clusterId: ClusterId(31), endpointId: EndpointNumber(0) },
@@ -457,7 +566,7 @@ const CHUNKED_ARRAY_WRITE_RESPONSE: WriteResponse = {
 };
 
 const INVOKE_COMMAND_REQUEST_WITH_EMPTY_ARGS: InvokeRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     timedRequest: false,
     invokeRequests: [
@@ -473,7 +582,7 @@ const INVOKE_COMMAND_REQUEST_WITH_EMPTY_ARGS: InvokeRequest = {
 };
 
 const INVOKE_COMMAND_REQUEST_TIMED_REQUIRED: InvokeRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     timedRequest: false,
     invokeRequests: [
@@ -489,7 +598,7 @@ const INVOKE_COMMAND_REQUEST_TIMED_REQUIRED: InvokeRequest = {
 };
 
 const INVOKE_COMMAND_RESPONSE_TIMED_REQUIRED: InvokeResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     invokeResponses: [
         {
@@ -502,7 +611,7 @@ const INVOKE_COMMAND_RESPONSE_TIMED_REQUIRED: InvokeResponse = {
 };
 
 const INVOKE_COMMAND_RESPONSE_TIMED_REQUIRED_SUCCESS: InvokeResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     invokeResponses: [
         {
@@ -515,7 +624,7 @@ const INVOKE_COMMAND_RESPONSE_TIMED_REQUIRED_SUCCESS: InvokeResponse = {
 };
 
 const INVOKE_COMMAND_REQUEST_WITH_NO_ARGS: InvokeRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     timedRequest: false,
     invokeRequests: [
@@ -526,7 +635,7 @@ const INVOKE_COMMAND_REQUEST_WITH_NO_ARGS: InvokeRequest = {
 };
 
 const INVOKE_COMMAND_REQUEST_MULTI: InvokeRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     timedRequest: false,
     invokeRequests: [
@@ -552,7 +661,7 @@ const INVOKE_COMMAND_REQUEST_MULTI: InvokeRequest = {
 };
 
 const INVOKE_COMMAND_REQUEST_INVALID: InvokeRequest = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     timedRequest: false,
     invokeRequests: [
@@ -563,7 +672,7 @@ const INVOKE_COMMAND_REQUEST_INVALID: InvokeRequest = {
 };
 
 const INVOKE_COMMAND_RESPONSE: InvokeResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     invokeResponses: [
         {
@@ -576,7 +685,7 @@ const INVOKE_COMMAND_RESPONSE: InvokeResponse = {
 };
 
 const INVOKE_COMMAND_RESPONSE_BUSY: InvokeResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     invokeResponses: [
         {
@@ -589,7 +698,7 @@ const INVOKE_COMMAND_RESPONSE_BUSY: InvokeResponse = {
 };
 
 const INVOKE_COMMAND_RESPONSE_INVALID: InvokeResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     invokeResponses: [
         {
@@ -602,7 +711,7 @@ const INVOKE_COMMAND_RESPONSE_INVALID: InvokeResponse = {
 };
 
 const INVOKE_COMMAND_RESPONSE_MULTI: InvokeResponse = {
-    interactionModelRevision: 10,
+    interactionModelRevision: 11,
     suppressResponse: false,
     invokeResponses: [
         {
@@ -756,8 +865,13 @@ describe("InteractionProtocol", () => {
         });
 
         it("replies with attribute values", async () => {
-            endpoint.addClusterServer(basicInfoClusterServer());
+            const basicInfoServer = basicInfoClusterServer();
+
+            endpoint.addClusterServer(basicInfoServer);
             interactionProtocol.setRootEndpoint(endpoint);
+
+            basicInfoServer.triggerStartUpEvent({ softwareVersion: 1 });
+            basicInfoServer.triggerStartUpEvent({ softwareVersion: 2 });
 
             const result = interactionProtocol.handleReadRequest(await getDummyMessageExchange(), READ_REQUEST);
 
@@ -765,8 +879,13 @@ describe("InteractionProtocol", () => {
         });
 
         it("replies with attribute values using (unused) version filter", async () => {
-            endpoint.addClusterServer(basicInfoClusterServer());
+            const basicInfoServer = basicInfoClusterServer();
+
+            endpoint.addClusterServer(basicInfoServer);
             interactionProtocol.setRootEndpoint(endpoint);
+
+            basicInfoServer.triggerStartUpEvent({ softwareVersion: 1 });
+            basicInfoServer.triggerStartUpEvent({ softwareVersion: 2 });
 
             const result = interactionProtocol.handleReadRequest(
                 await getDummyMessageExchange(),
@@ -777,8 +896,13 @@ describe("InteractionProtocol", () => {
         });
 
         it("replies with attribute values with active version filter", async () => {
-            endpoint.addClusterServer(basicInfoClusterServer());
+            const basicInfoServer = basicInfoClusterServer();
+
+            endpoint.addClusterServer(basicInfoServer);
             interactionProtocol.setRootEndpoint(endpoint);
+
+            basicInfoServer.triggerStartUpEvent({ softwareVersion: 1 });
+            basicInfoServer.triggerStartUpEvent({ softwareVersion: 2 });
 
             const result = interactionProtocol.handleReadRequest(
                 await getDummyMessageExchange(),